requirements.txt
torch
torchvision
numpy
pillow
matplotlib
scikit-learn
seaborn
tqdm
ipykernel
jupyter
